resolve merge conflicts

+++ ISHMS.Core/DTOs/Analytics/RiskBoardDto.cs
@@ -19,7 +19,7 @@ namespace ISHMS.Core.DTOs.Analytics
 
         public string RiskLevel { get; set; }
 
-        public int OxygenSaturation { get; set; }
+        public double OxygenSaturation { get; set; }
 
         public int HeartRate { get; set; }
